+++ packages/accel-record-core/src/fields.ts
@@ -1,258 +1,9 @@
-import { createId as cuid } from "@paralleldrive/cuid2";
-import { BaseDMMF, DMMF } from "prisma/prisma-client/runtime/library.js";
-import { Model, Models } from "./index.js";
+import { BaseDMMF } from "prisma/prisma-client/runtime/library.js";
+import { Model } from "./index.js";
+import { Association, buildAssociation } from "./model/association.js";
+import { Field } from "./model/field.js";
 
-/**
- * Represents an association between two models.
- */
-export class Association {
-  klass: string;
-  foreignKey: string;
-  primaryKey: string;
-  field: Field;
-  isBelongsTo: boolean;
-  through: string | undefined;
-
-  /**
-   * Creates a new Association instance.
-   * @param relation - The DMMF.Field object representing the relation.
-   * @param association - The Field object representing the association.
-   * @param primaryKeys - An array of primary key column names.
-   * @param foreignKeyColumns - An array of foreign key column names.
-   * @param primaryKeyColumns - An array of primary key column names.
-   */
-  constructor(
-    relation: DMMF.Field,
-    association: Field,
-    primaryKeys: readonly string[],
-    foreignKeyColumns: string[],
-    primaryKeyColumns: string[]
-  ) {
-    this.klass = association.type;
-    this.foreignKey = foreignKeyColumns[0] ?? "";
-    this.primaryKey = primaryKeyColumns[0] ?? "";
-    this.field = association;
-    if (this.foreignKey == "" && this.primaryKey == "") {
-      // Implicit many-to-many relations
-      this.isBelongsTo = false;
-      this.foreignKey = relation.type < association.type ? "A" : "B";
-      this.primaryKey = primaryKeys[0];
-      this.through = `_${association.relationName}`;
-    } else {
-      this.isBelongsTo = (relation.relationToFields?.length ?? 0) == 0;
-    }
-  }
-
-  /**
-   * Gets the model associated with this association.
-   */
-  get model() {
-    return Models[this.klass];
-  }
-
-  /**
-   * Checks if the association represents a has-one relation.
-   */
-  get isHasOne() {
-    return !this.field.isList && !this.isBelongsTo;
-  }
-
-  /**
-   * Gets the join key for the Implicit Many-to-Many association.
-   */
-  get joinKey() {
-    return this.foreignKey == "A" ? "B" : "A";
-  }
-}
-
-/**
- * Represents a field in a database table.
- */
-export class Field {
-  /**
-   * The name of the field.
-   */
-  name: string;
-
-  /**
-   * The name of the field in the database.
-   */
-  dbName: string;
-
-  /**
-   * The type of the field.
-   */
-  type: string;
-
-  /**
-   * The name of the related field in a relation, or null if not applicable.
-   */
-  relationName: string | null;
-
-  /**
-   * Indicates whether the field is a list.
-   */
-  isList: boolean;
-
-  /**
-   * Indicates whether the field is required.
-   */
-  isRequired: boolean;
-
-  /**
-   * The kind of the field.
-   */
-  kind: "scalar" | "object" | "enum" | "unsupported";
-
-  /**
-   * Indicates whether the field is an updated timestamp.
-   */
-  isUpdatedAt: boolean;
-
-  /**
-   * The default value of the field.
-   */
-  default: any;
-
-  /**
-   * The foreign keys associated with the field.
-   */
-  foreignKeys: string[];
-
-  /**
-   * The primary keys associated with the field.
-   */
-  primaryKeys: string[];
-
-  /**
-   * Constructs a new Field instance.
-   * @param field - The DMMF.Field object representing the field.
-   */
-  constructor(field: DMMF.Field) {
-    this.name = field.name;
-    this.dbName = field.dbName ?? field.name;
-    this.type = field.type;
-    this.relationName = field.relationName?.toString() ?? null;
-    this.isList = !!field.isList;
-    this.isRequired = !!field.isRequired;
-    this.kind = field.kind.toString() as
-      | "scalar"
-      | "object"
-      | "enum"
-      | "unsupported";
-    this.isUpdatedAt = !!field.isUpdatedAt;
-    this.default = field.default?.valueOf() ?? undefined;
-    this.foreignKeys = field.relationFromFields?.map((f) => f) ?? [];
-    this.primaryKeys = field.relationToFields?.map((f) => f) ?? [];
-  }
-
-  /**
-   * Checks if the default is "now" function.
-   * @returns True if the default value is "now", false otherwise.
-   */
-  get defaultIsNow() {
-    return this.default != undefined && this.default.name === "now";
-  }
-
-  /**
-   * Checks if the default is "uuid" function.
-   * @returns True if the default value is "uuid", false otherwise.
-   */
-  get defaultIsUuid() {
-    return this.default != undefined && this.default.name === "uuid";
-  }
-
-  /**
-   * Checks if the default is "cuid" function.
-   * @returns True if the default value is "cuid", false otherwise.
-   */
-  get defaultIsCuid() {
-    return this.default != undefined && this.default.name === "cuid";
-  }
-
-  /**
-   * Gets the scalar default value.
-   * @returns The scalar default value, or undefined if not applicable.
-   */
-  get scalarDefault() {
-    if (this.default != undefined && typeof this.default != "object") {
-      return this.cast(this.default);
-    }
-    return undefined;
-  }
-
-  /**
-   * Gets the initial value for the field.
-   * @returns The initial value for the field.
-   */
-  getInitialValue() {
-    if (this.defaultIsUuid) return crypto.randomUUID();
-    if (this.defaultIsCuid) return cuid();
-    return this.scalarDefault;
-  }
-
-  /**
-   * Casts a value to the appropriate type based on the field's type.
-   * @param value - The value to cast.
-   * @returns The casted value.
-   */
-  cast(value: any) {
-    if (value == undefined) return value;
-    switch (this.type) {
-      case "BigInt":
-      case "Decimal":
-      case "Float":
-      case "Int":
-        return Number(value);
-      case "Bytes":
-      case "String":
-        return String(value);
-      case "Boolean":
-        return !!value;
-      case "DateTime":
-        return new Date(value);
-      case "JSON":
-        return JSON.parse(value);
-      default:
-        return value;
-    }
-  }
-}
-
-const buildAssociation = (
-  model: DMMF.Model,
-  field: Field,
-  primaryKeys: readonly string[]
-) => {
-  const rModel = dmmf.datamodel.models.find((m) => m.name === field.type)!;
-  const r = rModel?.fields.find((f) => f.relationName === field.relationName);
-  if (r?.relationFromFields == undefined || r?.relationToFields == undefined)
-    return undefined;
-
-  const foreignKeys1 = r.relationFromFields.map(
-    (name) => rModel.fields.find((f) => f.name === name)?.dbName ?? name
-  );
-  const foreignKeys2 = field.foreignKeys.map(
-    (name) => model.fields.find((f) => f.name === name)?.dbName ?? name
-  );
-
-  const primaryKeys1 = r.relationToFields.map(
-    (name) => model.fields.find((f) => f.name === name)?.dbName ?? name
-  );
-  const primaryKeys2 = field.primaryKeys.map(
-    (name) => rModel.fields.find((f) => f.name === name)?.dbName ?? name
-  );
-
-  return new Association(
-    r,
-    field,
-    primaryKeys,
-    foreignKeys1.concat(foreignKeys2),
-    primaryKeys1.concat(primaryKeys2)
-  );
-};
-
-let dmmf: BaseDMMF;
+export let dmmf: BaseDMMF;
 
 export const loadDmmf = async () => {
   const { Prisma } = await import("@prisma/client");
