+++ packages/accel-record-core/src/migration/mysqlMigrator.ts
@@ -4,25 +4,7 @@ import { Migrator } from "./migrator.js";
 
 export class MySQLMigrator extends Migrator {
   async ensureDatabaseExists() {
-    const config = getConfig();
-    const parse = () => {
-      const knexConfig = getKnexConfig(config);
-      if (typeof knexConfig?.connection == "string") {
-        const u = new URL(knexConfig.connection);
-        const database = u.pathname.replace("/", "");
-        u.pathname = "";
-        const newConfig = { ...knexConfig, connection: u.toString() };
-        return [database, newConfig];
-      } else if (knexConfig) {
-        // @ts-ignore
-        const { database, ...rest } = knexConfig.connection;
-        const newConfig = { ...knexConfig, connection: rest };
-        return [database, newConfig];
-      } else {
-        throw new Error("Invalid knexConfig");
-      }
-    };
-    const [database, newConfig] = parse();
+    const [database, newConfig] = parseConfig();
     const knex = Knex(newConfig);
     const exists = await knex.raw(`SHOW DATABASES LIKE "${database}";`);
     if (exists[0].length == 0) {
@@ -36,6 +18,24 @@ export class MySQLMigrator extends Migrator {
     return this.knex.raw(CREATE_LOGS_TABLE_DDL);
   }
 }
+
+const parseConfig = () => {
+  const knexConfig = getKnexConfig(getConfig());
+  if (typeof knexConfig?.connection == "string") {
+    const u = new URL(knexConfig.connection);
+    const database = u.pathname.replace("/", "");
+    u.pathname = "";
+    const newConfig = { ...knexConfig, connection: u.toString() };
+    return [database, newConfig];
+  } else if (knexConfig) {
+    // @ts-ignore
+    const { database, ...rest } = knexConfig.connection;
+    const newConfig = { ...knexConfig, connection: rest };
+    return [database, newConfig];
+  } else {
+    throw new Error("Invalid knexConfig");
+  }
+};
 
 const CREATE_LOGS_TABLE_DDL = `CREATE TABLE IF NOT EXISTS \`_prisma_migrations\` (
   \`id\` varchar(36) COLLATE utf8mb4_unicode_ci NOT NULL,
