+++ packages/accel-record-core/src/migration/postgresqlMigrator.ts
@@ -4,28 +4,7 @@ import { Migrator } from "./migrator.js";
 
 export class PostgresqlMigrator extends Migrator {
   async ensureDatabaseExists() {
-    const config = getConfig();
-    const parse = () => {
-      const knexConfig = getKnexConfig(config);
-      if (typeof knexConfig?.connection == "string") {
-        const u = new URL(knexConfig.connection);
-        const database = u.pathname.replace("/", "");
-        u.pathname = "postgres";
-        const newConfig = { ...knexConfig, connection: u.toString() };
-        return [database, newConfig];
-      } else if (knexConfig) {
-        // @ts-ignore
-        const { database, ...rest } = knexConfig.connection;
-        const newConfig = {
-          ...knexConfig,
-          connection: { ...rest, database: "postgres" },
-        };
-        return [database, newConfig];
-      } else {
-        throw new Error("Invalid knexConfig");
-      }
-    };
-    const [database, newConfig] = parse();
+    const [database, newConfig] = parseConfig();
     const knex = Knex(newConfig);
     const exists = await knex.raw(
       `SELECT 1 FROM pg_database WHERE datname='${database}'`
@@ -41,6 +20,28 @@ export class PostgresqlMigrator extends Migrator {
     return this.knex.raw(CREATE_LOGS_TABLE_DDL);
   }
 }
+
+const parseConfig = () => {
+  const config = getConfig();
+  const knexConfig = getKnexConfig(config);
+  if (typeof knexConfig?.connection == "string") {
+    const u = new URL(knexConfig.connection);
+    const database = u.pathname.replace("/", "");
+    u.pathname = "postgres";
+    const newConfig = { ...knexConfig, connection: u.toString() };
+    return [database, newConfig];
+  } else if (knexConfig) {
+    // @ts-ignore
+    const { database, ...rest } = knexConfig.connection;
+    const newConfig = {
+      ...knexConfig,
+      connection: { ...rest, database: "postgres" },
+    };
+    return [database, newConfig];
+  } else {
+    throw new Error("Invalid knexConfig");
+  }
+};
 
 const CREATE_LOGS_TABLE_DDL = `CREATE TABLE IF NOT EXISTS public._prisma_migrations (
     id character varying(36) NOT NULL,
